requirements.txt
# Core RAG (Python 3.12 compatible)
langchain==0.3.1
langchain-community==0.3.1
langchain-google-genai==2.0.5
llama-index==0.12.14

# Embeddings & Vector Stores
chromadb==0.6.3
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1

# LLM Providers
google-generativeai==0.8.3
groq==0.13.0

# Document Processing
pymupdf==1.24.14
beautifulsoup4==4.12.3
unstructured==0.16.13
python-docx==1.1.2

# Retrieval & Reranking
rank-bm25==0.2.2
transformers==4.47.1
torch==2.5.1
bitsandbytes==0.45.0
accelerate==0.34.2

# Evaluation
ragas==0.2.10
bert-score==0.3.13
rouge-score==0.1.2

# API & Web
fastapi==0.115.6
uvicorn==0.34.0
streamlit==1.41.1
pydantic==2.10.4
pydantic-settings==2.7.0

# Monitoring & Logging
prometheus-client==0.21.1
python-json-logger==3.2.1

# Cache
redis==5.2.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
tqdm==4.67.1
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
